+++ pyhealth/interpret/tasks/__init__.py
@@ -1,3 +1,2 @@
-from pyhealth.interpret import methods 
 from pyhealth.interpret.tasks.medical_code import MedicalCodeInterpreter
 from pyhealth.interpret.tasks.vision import VisionInterpreter
